requirements.txt
requests
pandas
playwright
beautifulsoup4
python-dotenv
